Added other search engines
Other skills have the line **skills** before **-name** so I added it to keep all configuration examples the same.
I was unsure how to show the list of other engines but only keeping the default (Google) activated. In the end, I just commented the beginning of the line. Let me know if you would rather do this in some other way.